+++ src/basic/__tests__/basic.test.tsx
@@ -12,7 +12,7 @@ import { CartPage } from "../../refactoring/pages/CartPage";
 import { AdminPage } from "../../refactoring/pages/AdminPage";
 import { CartItem, Coupon, Product } from "../../types";
 import { useCart, useCoupons, useProducts } from "../../refactoring/hooks";
-import * as cartUtils from "../../refactoring/models/cart";
+import * as cartUtils from "../../refactoring/cart/cart";
 
 const mockProducts: Product[] = [
   {
