+++ src/refactoring/App.tsx
@@ -2,7 +2,8 @@ import { useState } from "react";
 import { CartPage } from "./pages/CartPage.tsx";
 import { AdminPage } from "./pages/AdminPage.tsx";
 import { Coupon, Product } from "../types.ts";
-import { useCoupons, useProducts } from "./hooks";
+import { useCoupons } from "./coupon/useCoupon.ts";
+import { useProducts } from "./product/useProduct.ts";
 
 const initialProducts: Product[] = [
   {
